fix bug where recognizing speech cannot be deleted (#132)
* fix bug where recognizing speech cannot be deleted

* remove audio file when removing recognized speech

* change variable names

+++ whisper-ios/Views/RecordList.swift
@@ -3,7 +3,7 @@ import SwiftUI
 struct RecordList: View {
     @Binding var recognizingSpeechIds: [UUID]
     @Binding var recognizedSpeeches: [RecognizedSpeech]
-    @Binding var isActives: [Bool]
+    @Binding var isRecordDetailActives: [Bool]
 
     var body: some View {
         VStack {
@@ -17,7 +17,7 @@ struct RecordList: View {
             RecognitionPane(
                 recognizingSpeechIds: $recognizingSpeechIds,
                 recognizedSpeeches: $recognizedSpeeches,
-                isActives: $isActives
+                isRecordDetailActives: $isRecordDetailActives
             )
         }
     }
@@ -46,7 +46,7 @@ struct RecordList: View {
                         recognizedSpeech: recognizedSpeech,
                         isRecognizing: recognizingSpeechIds.contains(recognizedSpeech.id)
                     )),
-                    isActive: $isActives[idx]
+                    isActive: $isRecordDetailActives[idx]
                 ) {
                     HStack {
                         Image(systemName: "mic.square.fill")
@@ -90,10 +90,25 @@ struct RecordList: View {
 
     private func deleteRecognizedSpeech(indexSet: IndexSet) {
         for i in indexSet {
-            CoreDataRepository.deleteRecognizedSpeech(recognizedSpeech: recognizedSpeeches[i])
+            // RecognizedSpeech is inserted to recognizedSpeeches array right after
+            // finishing recording, but saved to coredata after ASR is completed.
+            if recognizingSpeechIds.contains(recognizedSpeeches[i].id) {
+                recognizingSpeechIds.removeAll { id in id == recognizedSpeeches[i].id }
+            } else {
+                CoreDataRepository.deleteRecognizedSpeech(recognizedSpeech: recognizedSpeeches[i])
+            }
+
+            do {
+                // TODO: fix this (issue #25)
+                let fileName = recognizedSpeeches[i].audioFileURL.lastPathComponent
+                let url = getURLByName(fileName: fileName)
+                try FileManager.default.removeItem(at: url)
+            } catch {
+                Logger.error("Failed to remove audio file.")
+            }
+            recognizedSpeeches.remove(at: i)
+            isRecordDetailActives.remove(at: i)
         }
-        recognizedSpeeches.remove(atOffsets: indexSet)
-        isActives.remove(atOffsets: indexSet)
     }
 }
 
@@ -105,7 +120,7 @@ class RecordList_Previews: PreviewProvider {
             RecordList(
                 recognizingSpeechIds: .constant([]),
                 recognizedSpeeches: .constant(recognizedSpeechs),
-                isActives: .constant([Bool](repeating: false, count: recognizedSpeechs.count))
+                isRecordDetailActives: .constant([Bool](repeating: false, count: recognizedSpeechs.count))
             )
             .previewDevice(PreviewDevice(rawValue: "iPhone 14 Pro Max"))
             .previewDisplayName("iphone")
@@ -113,7 +128,7 @@ class RecordList_Previews: PreviewProvider {
             RecordList(
                 recognizingSpeechIds: .constant([]),
                 recognizedSpeeches: .constant(recognizedSpeechs),
-                isActives: .constant([Bool](repeating: false, count: recognizedSpeechs.count))
+                isRecordDetailActives: .constant([Bool](repeating: false, count: recognizedSpeechs.count))
             )
             .previewDevice(PreviewDevice(rawValue: "iPad Pro (12.9-inch) (4th generation)"))
             .previewDisplayName("ipad")
@@ -121,7 +136,7 @@ class RecordList_Previews: PreviewProvider {
             RecordList(
                 recognizingSpeechIds: .constant([]),
                 recognizedSpeeches: .constant([]),
-                isActives: .constant([])
+                isRecordDetailActives: .constant([])
             )
             .previewDevice(PreviewDevice(rawValue: "iPhone 14 Pro Max"))
             .previewDisplayName("iphone no record")
